CLM-1651 CLM-4385 Remove old user viewed notifiations on  cache update (PR #1135)

+++ insight-brain-data/src/test/java/com/sonatype/insight/brain/dataaccess/TemporaryEntity.java
@@ -222,7 +222,10 @@ public class TemporaryEntity
     }
 
     for (UserViewedProductNotification userViewedNotificationMapping : userViewedNotificationMappings) {
-      userViewedNotificationMappingDAO.delete(userViewedNotificationMapping);
+      if ((userViewedNotificationMapping = userViewedNotificationMappingDAO.getByUsernameAndNotificationId(
+          userViewedNotificationMapping.getUsername(), userViewedNotificationMapping.getNotificationId())) != null) {
+        userViewedNotificationMappingDAO.delete(userViewedNotificationMapping);
+      }
     }
   }
